requirements.txt
comtypes==1.1.10
mss~=6.1.0
numpy~=1.20.3
opencv-python~=4.5.2.52
Pillow~=8.2.0
pygame~=2.0.1
